Fix Vaccinator charge sounds playing 2+ times/tick
Due to prediction, the PrimaryAttack() function runs multiple times every tick.
Every time the function runs on a tick where the player gains another charge, it will stack a copy of the sound that plays, which quickly becomes extremely loud. Commonly this will play 2 or 3, but in testing I've seen upwards of 7 or 8 playing at the same time.